B6: stop truncating the fractional segregating-site count
segsites() returned int. When the SFS has been subsampled for missing data
its bin counts are fractional expectations, so S is fractional too, and the
truncation discarded up to a whole site -- from the reported S column and,
worse, from the value passed into Tajima's D as both numerator term and
variance term.

On the subsample-toy fixture the analytic S is 5.5 and divstats reported 5.
With S carried as a double, all three statistics now match the independent
calculation exactly:

                      pi          S          D
  analytic       2.707143   5.500000   0.701735
  divstats       2.70714    5.5        0.701735

An identical double-returning s_from_sfs() sat ten lines below segsites(),
never called and not declared in the header -- evidently the intended fix,
never wired up. Folded into segsites() rather than left as a second copy, so
both call sites in divstats-wintools.cpp pick it up without change.

Golden diff over the 22 cases, measured by running the previous commit's
binary and this one against identical inputs: 12 cases move, 10 are
unchanged. Only S and D columns move anywhere -- pi is untouched, which is
why sites-pi-only and ehh-part (which report pi but no S or D) are both
unchanged, along with all EHH and pik cases and missing-nosub, where without
subsampling S is already integral.

Maximum relative deviation per column:
  whole-window   S 10.0%    D 58.4%
  partition      S_A 29.2%   S_C 29.3%   D_A 63.0%   D_B 27.4%   D_C 33214%
D_C's figure is a window where Tajima's D sits very near zero, so the
relative change is enormous while the absolute change is small.

Also fixes a build hazard hit while making this change: the Makefile declared
no header dependencies, so editing a header left stale objects linked against
the old declarations. Changing segsites() from int to double with a stale
divstats-wintools.o produced a binary that read the return value out of the
integer register and printed S = 5.66e+07. Every object now depends on every
header -- over-broad, but always correct. Anyone who has built this tree
incrementally after a header edit should assume their binary was mislinked.

This changes published output and warrants a version bump before release.

See divstats_code_review.md finding B6.

+++ src/divstats-winstats.h
@@ -50,7 +50,7 @@ double fayWuH_from_sfs(array_t *sfs, double pi = -9);
 double ehh_from_hfs(HaplotypeFrequencySpectrum *hfs);
 double ehhk_from_hfs(HaplotypeFrequencySpectrum *hfs, int k);
 
-int segsites(array_t *sfs);
+double segsites(array_t *sfs);
 double calc_a1(int n);
 double calc_a2(int n);
 double calc_e1(int n, double a1);
